+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Only write entries that are worth mentioning to users.
 - Lib: Move `MCPTool` from `kimi_cli.tools.mcp` to `kimi_cli.soul.toolset`
 - LLM: Add provider type `vertexai` to support Vertex AI
 - LLM: Rename Gemini Developer API provider type from `google_genai` to `gemini`
+- Config: Migrate config file from JSON to TOML
 
 ## [0.65] - 2025-12-16